[components] Move dg cli back to verb-first (rearrange tests)
