--- python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_component_type_commands.py
@@ -1,376 +1,0 @@
-import shutil
-import textwrap
-from pathlib import Path
-
-from dagster_components.utils import format_error_message
-from dagster_dg.component import RemoteComponentRegistry
-from dagster_dg.component_key import GlobalComponentKey
-from dagster_dg.context import DgContext
-from dagster_dg.utils import ensure_dagster_dg_tests_import, set_toml_value
-
-ensure_dagster_dg_tests_import()
-
-from dagster_dg_tests.utils import (
-    ProxyRunner,
-    assert_runner_result,
-    fixed_panel_width,
-    isolated_components_venv,
-    isolated_example_component_library_foo_bar,
-    match_terminal_box_output,
-    modify_pyproject_toml,
-)
-
-# ########################
-# ##### SCAFFOLD
-# ########################
-
-
-def test_component_type_scaffold_success() -> None:
-    with (
-        ProxyRunner.test() as runner,
-        isolated_example_component_library_foo_bar(runner),
-    ):
-        result = runner.invoke("scaffold", "component-type", "baz")
-        assert_runner_result(result)
-        assert Path("foo_bar/lib/baz.py").exists()
-        dg_context = DgContext.from_config_file_discovery_and_cli_config(Path.cwd(), {})
-        registry = RemoteComponentRegistry.from_dg_context(dg_context)
-        assert registry.has_global(GlobalComponentKey(name="baz", namespace="foo_bar"))
-
-
-def test_component_type_scaffold_already_exists_fails() -> None:
-    with (
-        ProxyRunner.test() as runner,
-        isolated_example_component_library_foo_bar(runner),
-    ):
-        result = runner.invoke("scaffold", "component-type", "baz")
-        assert_runner_result(result)
-        result = runner.invoke("scaffold", "component-type", "baz")
-        assert_runner_result(result, exit_0=False)
-        assert "already exists" in result.output
-
-
-def test_component_type_scaffold_succeeds_non_default_component_lib_package() -> None:
-    with (
-        ProxyRunner.test() as runner,
-        isolated_example_component_library_foo_bar(runner, lib_package_name="foo_bar._lib"),
-    ):
-        result = runner.invoke(
-            "scaffold",
-            "component-type",
-            "baz",
-        )
-        assert_runner_result(result)
-        assert Path("foo_bar/_lib/baz.py").exists()
-        dg_context = DgContext.from_config_file_discovery_and_cli_config(Path.cwd(), {})
-        registry = RemoteComponentRegistry.from_dg_context(dg_context)
-        assert registry.has_global(GlobalComponentKey(name="baz", namespace="foo_bar"))
-
-
-def test_component_type_scaffold_fails_components_lib_package_does_not_exist() -> None:
-    with (
-        ProxyRunner.test() as runner,
-        isolated_example_component_library_foo_bar(runner),
-    ):
-        with modify_pyproject_toml() as toml:
-            set_toml_value(toml, ("tool", "dg", "component_lib_package"), "foo_bar._lib")
-        result = runner.invoke(
-            "scaffold",
-            "component-type",
-            "baz",
-        )
-        assert_runner_result(result, exit_0=False)
-        assert "Components lib package `foo_bar._lib` is not installed" in str(result.exception)
-
-
-# ########################
-# ##### DOCS
-# ########################
-
-
-def test_component_type_docs_success():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke("docs", "component-type", "simple_asset@dagster_components.test")
-        assert_runner_result(result)
-
-
-def test_component_type_docs_success_output_console():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke(
-            "docs",
-            "component-type",
-            "complex_schema_asset@dagster_components.test",
-            "--output",
-            "cli",
-        )
-        assert_runner_result(result)
-        assert "<html" in result.output
-        assert "An asset that has a complex params schema." in result.output
-
-
-# ########################
-# ##### INFO
-# ########################
-
-_EXPECTED_COMPONENT_TYPE_INFO_FULL = textwrap.dedent("""
-    simple_pipes_script_asset@dagster_components.test
-
-    Description:
-
-    A simple asset that runs a Python script with the Pipes subprocess client.
-
-    Because it is a pipes asset, no value is returned.
-
-    Scaffold params schema:
-
-    {
-        "properties": {
-            "asset_key": {
-                "title": "Asset Key",
-                "type": "string"
-            },
-            "filename": {
-                "title": "Filename",
-                "type": "string"
-            }
-        },
-        "required": [
-            "asset_key",
-            "filename"
-        ],
-        "title": "SimplePipesScriptAssetSchema",
-        "type": "object"
-    }
-
-    Component params schema:
-
-    {
-        "properties": {
-            "asset_key": {
-                "title": "Asset Key",
-                "type": "string"
-            },
-            "filename": {
-                "title": "Filename",
-                "type": "string"
-            }
-        },
-        "required": [
-            "asset_key",
-            "filename"
-        ],
-        "title": "SimplePipesScriptAssetSchema",
-        "type": "object"
-    }
-""").strip()
-
-
-def test_component_type_info_all_metadata_success():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "simple_pipes_script_asset@dagster_components.test",
-        )
-        assert_runner_result(result)
-        assert result.output.strip().endswith(_EXPECTED_COMPONENT_TYPE_INFO_FULL)
-
-
-def test_component_type_info_all_metadata_empty_success():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "all_metadata_empty_asset@dagster_components.test",
-        )
-        assert_runner_result(result)
-        assert result.output.strip().endswith(
-            textwrap.dedent("""
-                all_metadata_empty_asset@dagster_components.test
-            """).strip()
-        )
-
-
-def test_component_type_info_flag_fields_success():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "simple_pipes_script_asset@dagster_components.test",
-            "--description",
-        )
-        assert_runner_result(result)
-        assert result.output.strip().endswith(
-            textwrap.dedent("""
-            A simple asset that runs a Python script with the Pipes subprocess client.
-
-            Because it is a pipes asset, no value is returned.
-        """).strip()
-        )
-
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "simple_pipes_script_asset@dagster_components.test",
-            "--scaffold-params-schema",
-        )
-        assert_runner_result(result)
-        assert result.output.strip().endswith(
-            textwrap.dedent("""
-                {
-                    "properties": {
-                        "asset_key": {
-                            "title": "Asset Key",
-                            "type": "string"
-                        },
-                        "filename": {
-                            "title": "Filename",
-                            "type": "string"
-                        }
-                    },
-                    "required": [
-                        "asset_key",
-                        "filename"
-                    ],
-                    "title": "SimplePipesScriptAssetSchema",
-                    "type": "object"
-                }
-            """).strip()
-        )
-
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "simple_pipes_script_asset@dagster_components.test",
-            "--component-params-schema",
-        )
-        assert_runner_result(result)
-        assert result.output.strip().endswith(
-            textwrap.dedent("""
-                {
-                    "properties": {
-                        "asset_key": {
-                            "title": "Asset Key",
-                            "type": "string"
-                        },
-                        "filename": {
-                            "title": "Filename",
-                            "type": "string"
-                        }
-                    },
-                    "required": [
-                        "asset_key",
-                        "filename"
-                    ],
-                    "title": "SimplePipesScriptAssetSchema",
-                    "type": "object"
-                }
-            """).strip()
-        )
-
-
-def test_component_type_info_multiple_flags_fails() -> None:
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "simple_pipes_script_asset@dagster_components.test",
-            "--description",
-            "--scaffold-params-schema",
-        )
-        assert_runner_result(result, exit_0=False)
-        assert (
-            "Only one of --description, --scaffold-params-schema, and --component-params-schema can be specified."
-            in result.output
-        )
-
-
-def test_component_type_info_undefined_component_type_fails() -> None:
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke(
-            "info",
-            "component-type",
-            "fake@fake",
-        )
-        assert_runner_result(result, exit_0=False)
-        assert "No component type `fake@fake` is registered" in result.output
-
-
-# ########################
-# ##### LIST
-# ########################
-
-_EXPECTED_COMPONENT_TYPES = textwrap.dedent("""
-    ┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-    ┃ Component Type                                    ┃ Summary                                                          ┃
-    ┡━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┩
-    │ all_metadata_empty_asset@dagster_components.test  │                                                                  │
-    │ complex_schema_asset@dagster_components.test      │ An asset that has a complex params schema.                       │
-    │ simple_asset@dagster_components.test              │ A simple asset that returns a constant string value.             │
-    │ simple_pipes_script_asset@dagster_components.test │ A simple asset that runs a Python script with the Pipes          │
-    │                                                   │ subprocess client.                                               │
-    └───────────────────────────────────────────────────┴──────────────────────────────────────────────────────────────────┘
-""").strip()
-
-_EXPECTED_COMPONENT_TYPES_JSON = textwrap.dedent("""
-    [
-        {
-            "key": "all_metadata_empty_asset@dagster_components.test",
-            "summary": null
-        },
-        {
-            "key": "complex_schema_asset@dagster_components.test",
-            "summary": "An asset that has a complex params schema."
-        },
-        {
-            "key": "simple_asset@dagster_components.test",
-            "summary": "A simple asset that returns a constant string value."
-        },
-        {
-            "key": "simple_pipes_script_asset@dagster_components.test",
-            "summary": "A simple asset that runs a Python script with the Pipes subprocess client."
-        }
-    ]
-
-""").strip()
-
-
-def test_list_component_types_success():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        with fixed_panel_width(width=120):
-            result = runner.invoke("list", "component-type")
-            assert_runner_result(result)
-            # strip the first line of logging output
-            output = "\n".join(result.output.split("\n")[1:])
-            match_terminal_box_output(output.strip(), _EXPECTED_COMPONENT_TYPES)
-
-
-def test_component_type_list_json_succeeds():
-    with ProxyRunner.test() as runner, isolated_components_venv(runner):
-        result = runner.invoke("list", "component-type", "--json")
-        assert_runner_result(result)
-        # strip the first line of logging output
-        output = "\n".join(result.output.split("\n")[1:])
-        assert output.strip() == _EXPECTED_COMPONENT_TYPES_JSON
-
-
-# Need to use capfd here to capture stderr from the subprocess invoked by the `list component-type`
-# command. This subprocess inherits stderr from the parent process, for whatever reason `capsys` does
-# not work.
-def test_component_type_list_bad_entry_point_fails(capfd):
-    with ProxyRunner.test() as runner, isolated_example_component_library_foo_bar(runner):
-        # Delete the component lib package referenced by the entry point
-        shutil.rmtree("foo_bar/lib")
-
-        # Disable cache to force re-discovery of deleted entry point
-        result = runner.invoke("list", "component-type", "--disable-cache", "--json")
-        assert_runner_result(result, exit_0=False)
-
-        expected_error_message = format_error_message("""
-            An error occurred while executing a `dagster-components` command in the
-            Python environment
-        """)
-        assert expected_error_message in result.output
-
-        captured = capfd.readouterr()
-        assert "Error loading entry point `foo_bar` in group `dagster.components`." in captured.err
