+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_environment_validation.py
@@ -200,7 +200,7 @@ def test_no_deployment_or_code_location_failure(spec: CommandSpec) -> None:
 # ########################
 
 
-# `dg component scaffold` is special because global options have to be inserted before the
+# `dg scaffold component` is special because global options have to be inserted before the
 # subcommand name, instead of just at the end.
 def _add_global_cli_options(cli_args: tuple[str, ...], *global_opts: str) -> list[str]:
     if cli_args[:2] == ("scaffold", "component"):
